%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileVector.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{[.FileVector}
\alias{[.FileVector}
\alias{FileVector.[}
\alias{[,FileVector-method}

\title{Gets a subset of elements of a file vector}

\description{
 Gets a subset of elements of a file vector.
}

\usage{
\method{[}{FileVector}(this, i=NULL)
}

\arguments{
  \item{i}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} or element indices.}
}

\value{
 Returns a \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:000-LB-LT_-.FileVector]{*[<-}()}.
  For more information see \code{\link{FileVector}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
