%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileMatrix.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{[< -.FileMatrix}
\alias{[<-.FileMatrix}
\alias{FileMatrix.[<-}
\alias{[<-,FileMatrix-method}

\title{Assigns values to a subset of elements of a file matrix}

\description{
 Assigns values to a subset of elements of a file matrix.
}

\usage{
\method{[}{FileMatrix}(this, i, j) <- value
}

\arguments{
  \item{i}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} or row indices.}
  \item{j}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} or column indices.}
  \item{value}{Values to be assigned to the selected elements.}
}

\value{
 Returns itself.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:000-LB-.FileMatrix]{*[}()}.
  For more information see \code{\link{FileMatrix}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
