%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileVector.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{[< -.FileVector}
\alias{[<-.FileVector}
\alias{FileVector.[<-}
\alias{[<-,FileVector-method}

\title{Assigns values to a subset of elements of a file vector}

\description{
 Assigns values to a subset of elements of a file vector.
}

\usage{
\method{[}{FileVector}(this, i=NULL) <- value
}

\arguments{
  \item{i}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} or element indices.}
  \item{value}{Values to be assigned to the selected elements.}
}

\value{
 Returns itself.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:000-LB-.FileVector]{*[}()}.
  For more information see \code{\link{FileVector}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
