%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileMatrix.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.FileMatrix}
\alias{as.character.FileMatrix}
\alias{FileMatrix.as.character}
\alias{as.character,FileMatrix-method}

\title{Returns a short string describing the file matrix}

\description{
 Returns a short string describing the file matrix.
}

\usage{
\method{as.character}{FileMatrix}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{FileMatrix}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
