%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileMatrix.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.matrix.FileMatrix}
\alias{as.matrix.FileMatrix}
\alias{FileMatrix.as.matrix}
\alias{as.matrix.FileMatrix}
\alias{as.matrix,FileMatrix-method}


\title{Returns the elements of a file matrix as an \R matrix}

\description{
 Returns the elements of a file matrix as an \R matrix, that is, imported into memory (if possible).
}

\usage{\method{as.matrix}{FileMatrix}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{matrix}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FileMatrix}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
