%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.vector.AbstractFileArray}
\alias{as.vector.AbstractFileArray}
\alias{AbstractFileArray.as.vector}
\alias{as.vector.AbstractFileArray}
\alias{as.vector,AbstractFileArray-method}


\title{Returns the elements of a file array as an R vector}

\description{
 Returns the elements of a file array as an R vector, that is, imported into memory (if possible).
}

\usage{\method{as.vector}{AbstractFileArray}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:as.vector.AbstractFileArray]{*as.vector}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
