%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{dim.AbstractFileArray}
\alias{dim.AbstractFileArray}
\alias{AbstractFileArray.dim}
\alias{dim,AbstractFileArray-method}

\title{Gets the dimension of the file array}

\description{
 Gets the dimension of the file array.
}

\usage{
\method{dim}{AbstractFileArray}(x)
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}} of length two.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
