%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{flush.AbstractFileArray}
\alias{flush.AbstractFileArray}
\alias{AbstractFileArray.flush}
\alias{flush.AbstractFileArray}
\alias{flush,AbstractFileArray-method}


\title{Internal: Flushes the write buffer}

\description{
 Internal: Flushes the write buffer.
}

\usage{\method{flush}{AbstractFileArray}(con, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
