%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBasename.AbstractFileArray}
\alias{getBasename.AbstractFileArray}
\alias{AbstractFileArray.getBasename}
\alias{getBasename.AbstractFileArray}
\alias{getBasename,AbstractFileArray-method}


\title{Gets the basename (filename) of the data file}

\description{
 Gets the basename (filename) of the data file.
}

\usage{\method{getBasename}{AbstractFileArray}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:getPathname.AbstractFileArray]{*getPathname}()}.
  \code{\link[R.huge:getBasename.AbstractFileArray]{*getBasename}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
