%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBytesPerCell.AbstractFileArray}
\alias{getBytesPerCell.AbstractFileArray}
\alias{AbstractFileArray.getBytesPerCell}
\alias{getBytesPerCell.AbstractFileArray}
\alias{getBytesPerCell,AbstractFileArray-method}


\title{Gets the number of bytes per element in a file array}

\description{
 Gets the number of bytes per element in a file array.
}

\usage{\method{getBytesPerCell}{AbstractFileArray}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\details{
 In \R, an \code{\link[base]{integer}}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
