%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCloneNumber.AbstractFileArray}
\alias{getCloneNumber.AbstractFileArray}
\alias{AbstractFileArray.getCloneNumber}
\alias{getCloneNumber.AbstractFileArray}
\alias{getCloneNumber,AbstractFileArray-method}


\title{Gets the clone number of the file array}

\description{
 Gets the clone number of the file array.
}

\usage{\method{getCloneNumber}{AbstractFileArray}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
