%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getComments.AbstractFileArray}
\alias{getComments.AbstractFileArray}
\alias{AbstractFileArray.getComments}
\alias{getComments.AbstractFileArray}
\alias{getComments,AbstractFileArray-method}


\title{Sets the comments for this file array}

\description{
 Sets the comments for this file array.
}

\usage{\method{getComments}{AbstractFileArray}(this, ...)}

\arguments{
  \item{comments}{A \code{\link[base]{character}} \code{\link[base]{vector}} of new comments.}
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) the old comments.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:getComments.AbstractFileArray]{*getComments}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
