%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.AbstractFileArray}
\alias{getPath.AbstractFileArray}
\alias{AbstractFileArray.getPath}
\alias{getPath,AbstractFileArray-method}

\title{Gets the path (directory) where the data file lives}

\description{
 Gets the path (directory) where the data file lives.
}

\usage{
\method{getPath}{AbstractFileArray}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:getPathname.AbstractFileArray]{*getPathname}()}.
  \code{\link[R.huge:getBasename.AbstractFileArray]{*getBasename}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
