%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPathname.AbstractFileArray}
\alias{getPathname.AbstractFileArray}
\alias{AbstractFileArray.getPathname}
\alias{getPathname,AbstractFileArray-method}

\title{Gets the full pathname to the data file}

\description{
 Gets the full pathname to the data file.
}

\usage{
\method{getPathname}{AbstractFileArray}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:getPath.AbstractFileArray]{*getPath}()}.
  \code{\link[R.huge:getBasename.AbstractFileArray]{*getBasename}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
