%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSizeOfComments.AbstractFileArray}
\alias{getSizeOfComments.AbstractFileArray}
\alias{AbstractFileArray.getSizeOfComments}
\alias{getSizeOfComments.AbstractFileArray}
\alias{getSizeOfComments,AbstractFileArray-method}


\title{Gets the number of bytes the comments occupies}

\description{
 Gets the number of bytes the comments occupies.
}

\usage{\method{getSizeOfComments}{AbstractFileArray}(this, comments=this$header$comments, ...)}

\arguments{
  \item{comments}{The comments to be measured.}
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
