%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSizeOfData.AbstractFileArray}
\alias{getSizeOfData.AbstractFileArray}
\alias{AbstractFileArray.getSizeOfData}
\alias{getSizeOfData,AbstractFileArray-method}

\title{Gets the size of the data section in bytes}

\description{
 Gets the size of the data section in bytes.
}

\usage{
\method{getSizeOfData}{AbstractFileArray}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{double}} (since the largest \code{\link[base]{integer}} is only 2147483648-1).
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:getBytesPerCell.AbstractFileArray]{*getBytesPerCell}()}.
  \code{\link[R.huge:length.AbstractFileArray]{*length}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
