%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isOpen.AbstractFileArray}
\alias{isOpen.AbstractFileArray}
\alias{AbstractFileArray.isOpen}
\alias{isOpen.AbstractFileArray}
\alias{isOpen,AbstractFileArray-method}


\title{Checks whether the data file of the file array is open or not}

\description{
 Checks whether the data file of the file array is open or not.
}

\usage{\method{isOpen}{AbstractFileArray}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:open.AbstractFileArray]{*open}()}.
  \code{\link[R.huge:close.AbstractFileArray]{*close}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
