%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{length.AbstractFileArray}
\alias{length.AbstractFileArray}
\alias{AbstractFileArray.length}
\alias{length,AbstractFileArray-method}

\title{Gets the number of elements in a file array}

\description{
 Gets the number of elements in a file array.
}

\usage{
\method{length}{AbstractFileArray}(x)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{double}} (since the largest \code{\link[base]{integer}} is only 2147483648-1).
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:dim.AbstractFileArray]{*dim}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
