%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileVector.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{names.FileVector}
\alias{names.FileVector}
\alias{FileVector.names}
\alias{names,FileVector-method}

\title{Gets the element names of a file vector}

\description{
 Gets the element names of a file vector.
}

\usage{
\method{names}{FileVector}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{FileVector}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
