%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileMatrix.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{ncol.FileMatrix}
\alias{ncol.FileMatrix}
\alias{FileMatrix.ncol}
\alias{ncol.FileMatrix}
\alias{ncol,FileMatrix-method}


\title{Gets the number of columns of the matrix}

\description{
 Gets the number of columns of the matrix.
}

\usage{\method{ncol}{FileMatrix}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{double}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:dim.FileMatrix]{*dim}()}.
  For more information see \code{\link{FileMatrix}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
