%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{open.AbstractFileArray}
\alias{open.AbstractFileArray}
\alias{AbstractFileArray.open}
\alias{open.AbstractFileArray}
\alias{open,AbstractFileArray-method}


\title{Opens a connection to the data file of the file array}

\description{
 Opens a connection to the data file of the file array.
}

\usage{\method{open}{AbstractFileArray}(con, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns (invisibly) itself.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:isOpen.AbstractFileArray]{*isOpen}()}.
  \code{\link[R.huge:close.AbstractFileArray]{*close}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
