%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readHeader.AbstractFileArray}
\alias{readHeader.AbstractFileArray}
\alias{AbstractFileArray.readHeader}
\alias{readHeader.AbstractFileArray}
\alias{readHeader,AbstractFileArray-method}


\title{Read the header of a file array data file}

\description{
 Read the header of a file array data file.
}

\usage{\method{readHeader}{AbstractFileArray}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base]{list}} structure.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:writeHeader.AbstractFileArray]{*writeHeader}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
