%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setComments.AbstractFileArray}
\alias{setComments.AbstractFileArray}
\alias{AbstractFileArray.setComments}
\alias{setComments,AbstractFileArray-method}

\title{Sets the comments for this file array}

\description{
 Sets the comments for this file array.
}

\usage{
\method{setComments}{AbstractFileArray}(this, comments=NULL, ...)
}

\arguments{
  \item{comments}{A \code{\link[base]{character}} \code{\link[base]{vector}} of new comments.}
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) the old comments.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.huge:getComments.AbstractFileArray]{*getComments}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
