%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractFileArray.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeHeader.AbstractFileArray}
\alias{writeHeader.AbstractFileArray}
\alias{AbstractFileArray.writeHeader}
\alias{writeHeader.AbstractFileArray}
\alias{writeHeader,AbstractFileArray-method}


\title{Writes the header of a file array to file}

\description{
 Writes the header of a file array to file.
}

\usage{\method{writeHeader}{AbstractFileArray}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.huge:writeHeader.AbstractFileArray]{*writeHeader}()}.
  For more information see \code{\link{AbstractFileArray}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
